move some classes to common
